requirements.txt
asgiref
Django
psycopg2
sqlparse
tzdata
pillow
pytils
python-dotenv
djangorestframework
django-filter
